Push CellConsole logic down from Action
